backend/istak_backend/requirements.txt
# Django
Django==5.2.5

# Database
psycopg2-binary==2.9.8

# WSGI server
gunicorn==22.2.2

# REST API
djangorestframework==4.0.0

# CORS headers
django-cors-headers==4.3.0

# JWT authentication
djangorestframework-simplejwt==2.8.0

# Celery
celery==5.5.2
redis==5.2.0

# Optional: for development convenience
python-dotenv==1.0.0
